# stats: rewrite as native extension