--- a/source/extensions/filters/http/istio_stats/istio_stats.cc
+++ b/source/extensions/filters/http/istio_stats/istio_stats.cc
@@ -1,3 +1,17 @@
+// Copyright Istio Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 #include "source/extensions/filters/http/istio_stats/istio_stats.h"
 
 #include "envoy/registry/registry.h"
@@ -17,13 +31,23 @@ namespace IstioStats {
 namespace {
 constexpr absl::string_view CustomStatNamespace = "istiocustom";
 
-absl::string_view extractString(const envoy::config::core::v3::Node& node,
+absl::string_view extractString(const ProtobufWkt::Struct& metadata,
                                 const std::string& key) {
-  const auto& it = node.metadata().fields().find(key);
-  if (it == node.metadata().fields().end()) {
+  const auto& it = metadata.fields().find(key);
+  if (it == metadata.fields().end()) {
     return {};
   }
   return it->second.string_value();
+}
+
+absl::string_view extractMapString(const ProtobufWkt::Struct& metadata,
+                                   const std::string& map_key,
+                                   const std::string& key) {
+  const auto& it = metadata.fields().find(map_key);
+  if (it == metadata.fields().end()) {
+    return {};
+  }
+  return extractString(it->second.struct_value(), key);
 }
 
 // Process-wide context shared with all filter instances.
@@ -33,12 +57,18 @@ struct Context : public Singleton::Instance {
       : pool_(symbol_table),
         stat_namespace_(pool_.add(CustomStatNamespace)),
         requests_total_(pool_.add("istio_requests_total")),
+        request_duration_milliseconds_(
+            pool_.add("istio_request_duration_milliseconds")),
+        request_bytes_(pool_.add("istio_request_bytes")),
+        response_bytes_(pool_.add("istio_response_bytes")),
         unknown_(pool_.add("unknown")),
         source_(pool_.add("source")),
         destination_(pool_.add("destination")),
         latest_(pool_.add("latest")),
         http_(pool_.add("http")),
         grpc_(pool_.add("grpc")),
+        mtls_(pool_.add("mtls")),
+        none_(pool_.add("none")),
         reporter_(pool_.add("reporter")),
         source_workload_(pool_.add("source_workload")),
         source_workload_namespace_(pool_.add("source_workload_namespace")),
@@ -68,14 +98,24 @@ struct Context : public Singleton::Instance {
         connection_security_policy_(pool_.add("connection_security_policy")),
         response_code_(pool_.add("response_code")),
         grpc_response_status_(pool_.add("grpc_response_status")),
-        workload_name_(pool_.add(extractString(node, "WORKLOAD_NAME"))),
-        namespace_(pool_.add(extractString(node, "NAMESPACE"))) {}
+        workload_name_(
+            pool_.add(extractString(node.metadata(), "WORKLOAD_NAME"))),
+        namespace_(pool_.add(extractString(node.metadata(), "NAMESPACE"))),
+        canonical_name_(pool_.add(extractMapString(
+            node.metadata(), "LABELS", "service.istio.io/canonical-name"))),
+        canonical_revision_(pool_.add(extractMapString(
+            node.metadata(), "LABELS", "service.istio.io/canonical-revision"))),
+        cluster_name_(pool_.add(extractString(node.metadata(), "CLUSTER_ID"))) {
+  }
 
   Stats::StatNamePool pool_;
 
   // Metric names.
   const Stats::StatName stat_namespace_;
   const Stats::StatName requests_total_;
+  const Stats::StatName request_duration_milliseconds_;
+  const Stats::StatName request_bytes_;
+  const Stats::StatName response_bytes_;
 
   // Constant names.
   const Stats::StatName unknown_;
@@ -84,6 +124,8 @@ struct Context : public Singleton::Instance {
   const Stats::StatName latest_;
   const Stats::StatName http_;
   const Stats::StatName grpc_;
+  const Stats::StatName mtls_;
+  const Stats::StatName none_;
 
   // Dimension names.
   const Stats::StatName reporter_;
@@ -118,6 +160,9 @@ struct Context : public Singleton::Instance {
   // Per-process constants.
   const Stats::StatName workload_name_;
   const Stats::StatName namespace_;
+  const Stats::StatName canonical_name_;
+  const Stats::StatName canonical_revision_;
+  const Stats::StatName cluster_name_;
 };
 using ContextSharedPtr = std::shared_ptr<Context>;
 
@@ -204,38 +249,106 @@ class IstioStatsFilter : public Http::PassThroughFilter {
           baggage_result.value());
     }
     switch (config_->reporter()) {
-      case Reporter::ServerSidecar:
+      case Reporter::ServerSidecar: {
         tags_.push_back({context_.source_workload_namespace_,
                          peer ? config_->resolve(peer->namespaceName())
                               : context_.unknown_});
         tags_.push_back({context_.source_workload_,
                          peer ? config_->resolve(peer->workloadName())
                               : context_.unknown_});
+        tags_.push_back({context_.source_canonical_service_,
+                         peer ? config_->resolve(peer->canonicalName())
+                              : context_.unknown_});
+        tags_.push_back({context_.source_canonical_revision_,
+                         peer ? config_->resolve(peer->canonicalRevision())
+                              : context_.unknown_});
+        tags_.push_back(
+            {context_.source_cluster_,
+             peer ? config_->resolve(peer->clusterName()) : context_.unknown_});
         tags_.push_back(
             {context_.destination_workload_namespace_, context_.namespace_});
         tags_.push_back(
             {context_.destination_workload_, context_.workload_name_});
+        tags_.push_back({context_.destination_canonical_service_,
+                         context_.canonical_name_});
+        tags_.push_back({context_.destination_canonical_revision_,
+                         context_.canonical_revision_});
+        // Specific to reporter:
+        const auto& info = decoder_callbacks_->streamInfo();
+        const auto mtls =
+            info.downstreamAddressProvider().sslConnection() != nullptr &&
+            info.downstreamAddressProvider()
+                .sslConnection()
+                ->peerCertificatePresented();
+        tags_.push_back({context_.connection_security_policy_,
+                         mtls ? context_.mtls_ : context_.none_});
         break;
+      }
       case Reporter::ClientSidecar:
         tags_.push_back(
             {context_.source_workload_namespace_, context_.namespace_});
         tags_.push_back({context_.source_workload_, context_.workload_name_});
+        tags_.push_back(
+            {context_.source_canonical_service_, context_.canonical_name_});
+        tags_.push_back({context_.source_canonical_revision_,
+                         context_.canonical_revision_});
         tags_.push_back({context_.destination_workload_namespace_,
                          peer ? config_->resolve(peer->namespaceName())
                               : context_.unknown_});
         tags_.push_back({context_.destination_workload_,
                          peer ? config_->resolve(peer->workloadName())
                               : context_.unknown_});
+        tags_.push_back({context_.destination_canonical_service_,
+                         peer ? config_->resolve(peer->canonicalName())
+                              : context_.unknown_});
+        tags_.push_back({context_.destination_canonical_revision_,
+                         peer ? config_->resolve(peer->canonicalRevision())
+                              : context_.unknown_});
+        tags_.push_back(
+            {context_.destination_cluster_,
+             peer ? config_->resolve(peer->clusterName()) : context_.unknown_});
         break;
     }
+    // principal
+    // app, version
     return Http::FilterHeadersStatus::Continue;
   }
 
   void onStreamComplete() override {
+    const auto& info = decoder_callbacks_->streamInfo();
+
+    // TODO: copy Http::CodeStatsImpl version for status codes and flags.
+    tags_.push_back(
+        {context_.response_code_,
+         config_->resolve(absl::StrCat(info.responseCode().value_or(0)))});
+    tags_.push_back(
+        {context_.response_flags_,
+         config_->resolve(StreamInfo::ResponseFlagUtils::toShortString(info))});
+
     Stats::Utility::counterFromElements(
         config_->scope_, {context_.stat_namespace_, context_.requests_total_},
         tags_)
         .inc();
+    auto duration = info.requestComplete();
+    if (duration.has_value()) {
+      Stats::Utility::histogramFromElements(
+          config_->scope_,
+          {context_.stat_namespace_, context_.request_duration_milliseconds_},
+          Stats::Histogram::Unit::Milliseconds, tags_)
+          .recordValue(absl::FromChrono(duration.value()) /
+                       absl::Milliseconds(1));
+    }
+    auto meter = info.getDownstreamBytesMeter();
+    if (meter) {
+      Stats::Utility::histogramFromElements(
+          config_->scope_, {context_.stat_namespace_, context_.request_bytes_},
+          Stats::Histogram::Unit::Bytes, tags_)
+          .recordValue(meter->wireBytesReceived());
+      Stats::Utility::histogramFromElements(
+          config_->scope_, {context_.stat_namespace_, context_.response_bytes_},
+          Stats::Histogram::Unit::Bytes, tags_)
+          .recordValue(meter->wireBytesSent());
+    }
   }
 
  private: