… src/envoy/BUILD
… test/envoye2e/driver/stats.go
… test/envoye2e/inventory.go
… test/envoye2e/stats_plugin/stats_test.go
… test/envoye2e/tcp_metadata_exchange/tcp_metadata_exchange_test.go
… testdata/filters/client_stats_network_filter.yaml.tmpl
… testdata/filters/extension_config_inbound.yaml.tmpl
… testdata/filters/extension_config_outbound.yaml.tmpl
… testdata/filters/server_stats_network_filter.yaml.tmpl
… testdata/filters/stats_inbound.yaml.tmpl
… testdata/filters/stats_outbound.yaml.tmpl
… testdata/metric/client_disable_host_header_fallback.yaml.tmpl
… testdata/metric/client_request_total.yaml.tmpl
… testdata/metric/client_request_total_endpoint_labels.yaml.tmpl
… testdata/metric/host_header_fallback.yaml.tmpl
… testdata/stats/client_config.yaml
… testdata/stats/client_config_disable_header_fallback.yaml