Signed-off-by: Kuat Yessenov <kuat@google.com>

This is a quick prototype of Istio stats extension written as a native extension. The advantages are numerous:

* Unblocks ability to implement metric expiry, memory accounting for multi tenancy, and opens paths for Envoy improvements in that area;
* Removes requirement to rely on regex parsing to strip tags. Tags are produced directly.
* Removes dependency on Wasm. 
* Removes dependency on flatbuffers.
* CPU improvements. It is well known that Istio proxy is not frugal with CPU. This PR removes a lot of abstraction overhead (via Wasm, flatbuffers, regexes), and can be made extremely efficient with regards to memory / CPU cycles (Envoy uses highly efficient packed strings which are sub-pointer size).

*Update*:
This extension conforms to the original extensions for the common case. The main gaps are:
* support for configurable dimensions and metrics via CEL;
* support for gRPC streaming metrics;
* some flatbuffers dependencies are still present to allow seamless transition.

Neither of those are difficult in principle to add, but this PR is sufficient for the purpose of rebasing for Ambient.

*Update 2*:
Added limited CEL support and removal/modification of metrics. I am not sure how far we really want to support that in the end, but removal/little modifications are fairly safe.