@zirain 2022-09-26
@kyessenov do you plan to land this in 1.16?
another question is can user use this in previous version(1.15 etc)?