@kyessenov 2022-09-26
I might be able to get this in 1.16 as an option. No, I'm not planning to backport to previous versions. I think this is mostly need for Ambient, since Wasm is hard to make multi-tenant and shared proxies are, well, shared.