@douglas-reid 2022-09-26
Of all of the benefits:

> Removes requirement to rely on regex parsing to strip tags. Tags are produced directly.

would probably be the biggest UX improvement for Istio customers. Dynamic ability to add tags would reduce so much pain (and reduce the learning curve).