@kyessenov 2022-09-27
> the new filter can run in previous envoy?

I think the internal stats APIs have not changed much for awhile, so yes, you could backport it to older Envoy releases.