@kyessenov 2022-10-07
Benchmark runs:
- [stats-old.pdf](https://github.com/istio/proxy/files/9737341/stats-old.pdf)
- [stats-new.pdf](https://github.com/istio/proxy/files/9737340/stats-new.pdf)
(load test at 3kqps for 30s for single threaded envoy pinned to a core, no peer info)

Stats overhead:
- before: PluginRootContext::report 0.03s (0.49%) of 1.17s (19.02%)
- after: HttpFilters::IstioStats 0.06s (1.37%) of 0.55s (12.53%)

So we save about 40% CPU, but could do better if we reduce per-request overhead further.