@istio-testing 2022-10-11
@kyessenov: The following test **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
test_proxy | eca7632b7ace59bdf1436357c6edb4f21c46ba26 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_proxy/4079/test_proxy/1579941863144558592) | true | `/test test`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->