@ymesika 2022-10-12
@kyessenov I'm trying to use this with Ambient's ztunnel and added it both as http and filter listener of the inbound.
While I see the `counterFromElements` being called with elements and tags I don't see anything when getting the metrics from the `:15020/stats`.
Is there some extra configuration that needs to be set? Not providing any configuration to the filters atm.