@ymesika 2022-10-12
> @ymesika Maybe you need to call /stats/prometheus to see the tagged metrics?

Sorry, that's what I called.. Copy-paste got truncated