@kyessenov 2022-10-12
@ymesika Not sure, maybe something to do with stats_config allow-list in the bootstrap? Try to clear it and see if it shows up.