@istio-testing 2022-10-12
@kyessenov: The following test **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
test_proxy | 46ba278749dc0874babc2cbb5dc3562d71e4b8d6 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_proxy/4079/test_proxy/1580288840118046720) | true | `/test test`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->